+++ mosip-packet-creator/src/main/java/io/mosip/testrig/dslrig/packetcreator/service/APIRequestUtil.java
@@ -106,11 +106,12 @@ public class APIRequestUtil {
 	}
 
     /**
-     * @deprecated Prefer {@link #clearRunScopedCache(String)} for a single DSL context.
+     * @deprecated Use {@link #clearRunScopedCache(String)} — global token clearing is unsafe under parallel runs.
      */
     @Deprecated
     public void clearToken() {
-    	AuthTokenStore.clearAll();
+    	throw new UnsupportedOperationException(
+    			"Global clearToken() clears all contexts; use clearRunScopedCache(contextKey) instead.");
     }
 
     public void clearRunScopedCache(String contextKey) {
