+++ mosip-packet-creator/src/main/resources/application.properties
@@ -15,7 +15,6 @@ mosip.test.regclient.clientid =
 mosip.test.regclient.secretkey=
 mosip.test.regclient.appId =
 mosip.test.primary.langcode=eng
-mosip.test.persona.Angulipath=
 ##################################################################################### 
 
 mosip.test.persona.configpath=./resource/
