+++ mosipTestDataProvider/resources/application.properties
@@ -95,7 +95,6 @@ mosip.test.schemaManager.configpath=../deploy/resource/
 mosip.test.policyManager.configpath=../deploy/resource/
 mosip.test.partnerManager.configpath=../deploy/resource/
 mosip.test.prereg.mapfile=..\\deploy\\Preregistration.properties
-mosip.test.persona.Angulipath=..\\deploy\\anguli\\
 mosip.test.temp=/temp/
 mosip.test.testcase.propertypath=..\\deploy\\testcases.properties
 mosip.test.post2slack=true
