+++ mosipTestDataProvider/src/io/mosip/testrig/dslrig/dataprovider/BiometricDataProvider.java
@@ -1,1809 +1,109 @@
 package io.mosip.testrig.dslrig.dataprovider;
 
-import java.awt.image.BufferedImage;
-import java.io.BufferedInputStream;
-import java.io.BufferedReader;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.FileFilter;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Enumeration;
-import java.util.HashMap;
 import java.util.Hashtable;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Random;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-import javax.imageio.ImageIO;
 import javax.xml.parsers.FactoryConfigurationError;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 
-import org.apache.commons.codec.binary.Hex;
-import org.apache.commons.validator.Var;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import io.mosip.testrig.dslrig.dataprovider.mds.MDSClient;
-import io.mosip.testrig.dslrig.dataprovider.mds.MDSClientInterface;
-import io.mosip.testrig.dslrig.dataprovider.mds.MDSClientNoMDS;
+import io.mosip.testrig.dslrig.dataprovider.biometric.CbeffAssembler;
+import io.mosip.testrig.dslrig.dataprovider.biometric.CbeffBirBuilder;
+import io.mosip.testrig.dslrig.dataprovider.biometric.FileBiometricLoader;
+import io.mosip.testrig.dslrig.dataprovider.biometric.MdsCaptureService;
 import io.mosip.testrig.dslrig.dataprovider.models.BioModality;
 import io.mosip.testrig.dslrig.dataprovider.models.BiometricDataModel;
 import io.mosip.testrig.dslrig.dataprovider.models.IrisDataModel;
 import io.mosip.testrig.dslrig.dataprovider.models.ResidentModel;
-import io.mosip.testrig.dslrig.dataprovider.models.mds.MDSDevice;
-import io.mosip.testrig.dslrig.dataprovider.models.mds.MDSDeviceCaptureModel;
 import io.mosip.testrig.dslrig.dataprovider.models.mds.MDSRCaptureModel;
-import io.mosip.testrig.dslrig.dataprovider.test.registrationclient.RegistrationSteps;
-import io.mosip.testrig.dslrig.dataprovider.util.CommonUtil;
-import io.mosip.testrig.dslrig.dataprovider.util.DataProviderConstants;
-import io.mosip.testrig.dslrig.dataprovider.util.FPClassDistribution;
-import io.mosip.testrig.dslrig.dataprovider.util.RestClient;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.xml.sax.SAXException;
-
-import com.jamesmurty.utils.XMLBuilder;
-import io.mosip.mock.sbi.test.CentralizedMockSBI;
-import io.mosip.testrig.dslrig.dataprovider.variables.VariableManager;
 
 public class BiometricDataProvider {
 
-	public static HashMap<String, Integer> portmap = new HashMap<>();
 	private static final Logger logger = LoggerFactory.getLogger(BiometricDataProvider.class);
 
-
-	private static final String XMLNS = "xmlns";
-	private static final String XMLNS_URL = "http://standards.iso.org/iso-iec/19785/-3/ed-2/";
-	private static final String MAJOR = "Major";
-	private static final String MINOR = "Minor";
-	private static final String CBEFFVERSION = "CBEFFVersion";
-	private static final String VERSION = "Version";
-	private static final String FALSE = "false";
-	private static final String BDBINFO = "BDBInfo";
-	private static final String BIRINFO = "BIRInfo";
-	private static final String INTEGRITY = "Integrity";
-	private static final String FORMAT = "Format";
-	private static final String CREATIONDATE = "CreationDate";
-	private static final String ORGANIZATION = "Organization";
-	private static final String MOSIP = "Mosip";
-	private static final String SUBTYPE = "Subtype";
-	private static final String PURPOSE = "Purpose";
-	private static final String LEVEL = "Level";
-	private static final String SHA_256 = "SHA-256";
-	private static final String ENROLL = "Enroll";
-	private static final String QUALITY = "Quality";
-	private static final String ALGORITHM = "Algorithm";
-	private static final String SCORE = "Score";
-	private static final String EXCEPTION = "EXCEPTION";
-	private static final String OTHERS = "others";
-	private static final String ENTRY = "entry";
-	private static final String RETRIES = "RETRIES";
-	private static final String SDK_SCORE = "SDK_SCORE";
-	private static final String FORCE_CAPTURED = "FORCE_CAPTURED";
-	private static final String PAYLOAD = "PAYLOAD";
-	private static final String SPEC_VERSION = "SPEC_VERSION";
-	public static final String AUTHCERTSPATH = "authCertsPath";
 	private static final String LEFTEYE = "leftEye";
 	private static final String RIGHTEYE = "rightEye";
-	private static final String RIGHT = "Right";
-	private static final String DIRPATH = "dirPath ";
-	private static final String SCENARIO = "scenario";
+	private static final String PAYLOAD = "PAYLOAD";
+	private static final String FALSE = "false";
 
-	static String buildBirIris(String irisInfo, String irisName, String jtwSign, String payload, String qualityScore,
-			boolean genarateValidCbeff, String exception, String contextKey) throws ParserConfigurationException,
-			FactoryConfigurationError, TransformerException, FileNotFoundException {
-		String today = CommonUtil.getUTCDateTime(null);
-		XMLBuilder builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL)
-				.e(VERSION).e(MAJOR).t("1").up().e(MINOR).t("1").up().up().e(CBEFFVERSION).e(MAJOR).t("1").up().e(MINOR)
-				.t("1").up().up().e(BIRINFO).e(INTEGRITY).t(FALSE).up().up().e(BDBINFO).e(FORMAT).e(ORGANIZATION)
-				.t(MOSIP).up().e("Type").t("9").up().up().e(CREATIONDATE).t(today).up().e("Type").t("Iris").up()
-				.e(SUBTYPE).t(irisName).up().e(LEVEL).t("Raw").up().e(PURPOSE).t(ENROLL).up().e(QUALITY).e(ALGORITHM)
-				.e(ORGANIZATION).t("HMAC").up().e("Type").t(SHA_256).up().up().e(SCORE)
-				.t((int) Math.round(Double.parseDouble(qualityScore)) + "").up().up().up()
-				.e("BDB").t(irisInfo).up().up();
-		if (jtwSign != null && payload != null) {
-			jtwSign = Base64.getEncoder().encodeToString(jtwSign.getBytes());
-			builder.e("SB").t(jtwSign).up().
-
-					e(OTHERS).e(ENTRY).a("key", EXCEPTION).t(exception).up().e(ENTRY).a("key", RETRIES).t("1").up()
-					.e(ENTRY).a("key", SDK_SCORE).t("0.0").up().e(ENTRY).a("key", FORCE_CAPTURED).t(FALSE).up().e(ENTRY)
-					.a("key", PAYLOAD).t(payload).up().e(ENTRY).a("key", SPEC_VERSION).t("0.9.5").up().up();
-		}
-		return builder.asString(null);
-	}
-
-	static String buildBirFinger(String fingerInfo, String fingerName, String jtwSign, String payload,
-			String qualityScore, boolean generateValidCbeff, String exception, String contextKey)
-			throws ParserConfigurationException, FactoryConfigurationError, TransformerException,
-			FileNotFoundException {
-		String today = CommonUtil.getUTCDateTime(null);
-		XMLBuilder builder = null;
-		String bdbKey = "BDB";
-		if (generateValidCbeff == false)
-			bdbKey = "invalidBDB";
-		builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL).e(VERSION)
-				.e(MAJOR).t("1").up().e(MINOR).t("1").up().up().e(CBEFFVERSION).e(MAJOR).t("1").up().e(MINOR).t("1")
-				.up().up().e(BIRINFO).e(INTEGRITY).t(FALSE).up().up().e(BDBINFO).e(FORMAT).e(ORGANIZATION).t(MOSIP).up()
-				.e("Type").t("7").up().up().e(CREATIONDATE).t(today).up().e("Type").t("Finger").up().e(SUBTYPE)
-				.t(fingerName).up().e(LEVEL).t("Raw").up().e(PURPOSE).t(ENROLL).up().e(QUALITY).e(ALGORITHM)
-				.e(ORGANIZATION).t("HMAC").up().e("Type").t(SHA_256).up().up().e(SCORE)
-				.t((int) Math.round(Double.parseDouble(qualityScore)) + "").up().up().up()
-				.e(bdbKey).t(fingerInfo).up().up();
-		if (jtwSign != null && payload != null) {
-			jtwSign = Base64.getEncoder().encodeToString(jtwSign.getBytes());
-			builder.e("SB").t(jtwSign).up().
-
-					e(OTHERS).e(ENTRY).a("key", EXCEPTION).t(exception).up().e(ENTRY).a("key", RETRIES).t("1").up()
-					.e(ENTRY).a("key", SDK_SCORE).t("0.0").up().e(ENTRY).a("key", FORCE_CAPTURED).t(FALSE).up().e(ENTRY)
-					.a("key", PAYLOAD).t(payload).up().e(ENTRY).a("key", SPEC_VERSION).t("0.9.5").up().up();
-		}
-		return builder.asString(null);
-	}
-
-	static String buildBirFace(String faceInfo, String jtwSign, String payload, String qualityScore,
-			boolean genarateValidCbeff, String exception, String contextKey) throws ParserConfigurationException,
-			FactoryConfigurationError, TransformerException, FileNotFoundException {
-		String today = CommonUtil.getUTCDateTime(null);
-		XMLBuilder builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL)
-				.e(VERSION).e(MAJOR).t("1").up().e(MINOR).t("1").up().up().e(CBEFFVERSION).e(MAJOR).t("1").up().e(MINOR)
-				.t("1").up().up().e(BIRINFO).e(INTEGRITY).t(FALSE).up().up().e(BDBINFO).e(FORMAT).e(ORGANIZATION)
-				.t(MOSIP).up().e("Type").t("8").up().up().e(CREATIONDATE).t(today).up().e("Type").t("Face").up()
-				.e(SUBTYPE).t("").up().e(LEVEL).t("Raw").up().e(PURPOSE).t(ENROLL).up().e(QUALITY).e(ALGORITHM)
-				.e(ORGANIZATION).t("HMAC").up().e("Type").t(SHA_256).up().up().e(SCORE)
-				.t((int) Math.round(Double.parseDouble(qualityScore)) + "").up().up().up()
-				.e("BDB").t(faceInfo).up().up();
-		if (jtwSign != null && payload != null) {
-			jtwSign = Base64.getEncoder().encodeToString(jtwSign.getBytes());
-			builder.e("SB").t(jtwSign).up().
-
-					e(OTHERS).e(ENTRY).a("key", EXCEPTION).t(exception).up().e(ENTRY).a("key", RETRIES).t("1").up()
-					.e(ENTRY).a("key", SDK_SCORE).t("0.0").up().e(ENTRY).a("key", FORCE_CAPTURED).t(FALSE).up().e(ENTRY)
-					.a("key", PAYLOAD).t(payload).up().e(ENTRY).a("key", SPEC_VERSION).t("0.9.5").up().up();
-
-		}
-		if (Double.parseDouble(qualityScore) >= 80)
-			VariableManager.setVariableValue(contextKey, "Biometric_Quality-Face", "level-9");
-		else
-			VariableManager.setVariableValue(contextKey, "Biometric_Quality-Face", "level-2");
-		return builder.asString(null);
-	}
-
-	static String buildBirExceptionPhoto(String faceInfo, String jtwSign, String payload, String qualityScore,
-			boolean genarateValidCbeff, String exception, String contextKey) throws ParserConfigurationException,
-			FactoryConfigurationError, TransformerException, FileNotFoundException {
-		String today = CommonUtil.getUTCDateTime(null);
-		XMLBuilder builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL)
-				.e(VERSION).e(MAJOR).t("1").up().e(MINOR).t("1").up().up().e(CBEFFVERSION).e(MAJOR).t("1").up().e(MINOR)
-				.t("1").up().up().e(BIRINFO).e(INTEGRITY).t(FALSE).up().up().e(BDBINFO).e(FORMAT).e(ORGANIZATION)
-				.t(MOSIP).up().e("Type").t("8").up().up().e(CREATIONDATE).t(today).up().e("Type").t("ExceptionPhoto")
-				.up().e(SUBTYPE).t("").up().e(LEVEL).t("Raw").up().e(PURPOSE).t(ENROLL).up().e(QUALITY).e(ALGORITHM)
-				.e(ORGANIZATION).t("HMAC").up().e("Type").t(SHA_256).up().up().e(SCORE)
-				.t((int) Math.round(Double.parseDouble(qualityScore)) + "").up().up().up()
-				.e("BDB").t(faceInfo).up().up();
-		if (jtwSign != null && payload != null) {
-			jtwSign = Base64.getEncoder().encodeToString(jtwSign.getBytes());
-			builder.e("SB").t(jtwSign).up().
-
-					e(OTHERS).e(ENTRY).a("key", EXCEPTION).t(exception).up().e(ENTRY).a("key", RETRIES).t("1").up()
-					.e(ENTRY).a("key", SDK_SCORE).t("0.0").up().e(ENTRY).a("key", FORCE_CAPTURED).t(FALSE).up().e(ENTRY)
-					.a("key", PAYLOAD).t(payload).up().e(ENTRY).a("key", SPEC_VERSION).t("0.9.5").up().up();
-
-		}
-
-
-		return builder.asString(null);
-	}
+	public static final String AUTHCERTSPATH = "authCertsPath";
 
 	public static List<BioModality> getModalitiesByType(List<BioModality> bioExceptions, String type) {
-		List<BioModality> lst = new ArrayList<BioModality>();
-
-		for (BioModality m : bioExceptions) {
-			if (m.getType().equalsIgnoreCase(type)) {
-				lst.add(m);
-			}
-		}
-		return lst;
+		return MdsCaptureService.getModalitiesByType(bioExceptions, type);
 	}
 
 	public static MDSRCaptureModel regenBiometricViaMDS(ResidentModel resident, String contextKey, String purpose,
 			String qualityScore, String process) throws Exception {
-		CentralizedMockSBI.stopSBI(contextKey);
-		BiometricDataModel biodata = null;
-		MDSRCaptureModel capture = null;
-
-		MDSClientInterface mds = null;
-		String val;
-		boolean bNoMDS = true;
-		String mdsprofilePath = null;
-		String profileName = null;
-		int port = 0;
-		val = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "mdsbypass").toString();
-		List<String> filteredAttribs = resident.getFilteredBioAttribtures();
-		List<BioModality> bioExceptions = resident.getBioExceptions();
-		List<String> bioexceptionlist = new ArrayList<String>();
-		try {
-			try {
-			if (val == null || val.equals("") || val.equals(FALSE)) {
-
-				val = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "mdsport").toString();
-				mdsprofilePath = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "mdsprofilepath")
-						.toString();
-
-
-				Path p12path = null;
-				boolean invalidCertFlag = Boolean
-						.parseBoolean(VariableManager.getVariableValue(contextKey, "invalidCertFlag").toString());
-
-				if (invalidCertFlag)
-					p12path = Paths.get(
-							VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "invalidCertpath").toString());
-				else
-					p12path = Paths.get(System.getProperty("java.io.tmpdir"),
-							VariableManager.getVariableValue(contextKey, "db-server").toString());
-
-				RestClient.logInfo(contextKey, "p12path" + p12path);
-
-				int maxLoopCount = Integer.parseInt(
-						VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "mdsPortLoopCount").toString());
-
-				while (maxLoopCount > 0) {
-					try {
-						port = CentralizedMockSBI.startSBI(contextKey, "Registration", "Biometric Device",
-								p12path.toString());
-					} catch (Exception e) {
-						logger.error("Exception occured during startSBI " + contextKey, e);
-					}
-					if (port != 0) {
-						RestClient.logInfo(contextKey, "Found the port " + contextKey + " port number is: " + port);
-						break;
-					}
-
-					maxLoopCount--;
-				}
-
-				if (port == 0) {
-					logger.error("Unable to find the port " + contextKey + " port number is: " + port);
-					return null;
-				}
-
-
-				portmap.put("port_" + contextKey, port);
-
-				mds = new MDSClient(port);
-
-				profileName = "res" + resident.getId();
-
-				if (resident.getBioExceptions() != null && !resident.getBioExceptions().isEmpty()) {
-					copyExceptionPhoto(mdsprofilePath, "Default", profileName);
-				}
-
-				if (process.equalsIgnoreCase("Update")) {
-
-					mds.updateProfile(mdsprofilePath, profileName, resident, contextKey, purpose);
-
-				} else {
-
-					mds.createProfile(mdsprofilePath, profileName, resident, contextKey, purpose);
-				}
-				mds.setProfile(profileName, port, contextKey);
-
-			} else {
-				mds = new MDSClientNoMDS();
-				bNoMDS = false;
-				profileName = "res" + resident.getId();
-				mds.createProfile(mdsprofilePath, profileName, resident, contextKey, purpose);
-				mds.setProfile(profileName, port, contextKey);
-			}
-
-
-			HashMap<String, Integer> portAsPerKey = BiometricDataProvider.portmap;
-			RegistrationSteps steps = new RegistrationSteps();
-			steps.setMDSscore(portAsPerKey.get("port_" + contextKey), "Biometric Device", qualityScore, contextKey);
-			RestClient.logInfo(contextKey, "mds score is changed to : " + qualityScore);
-			biodata = resident.getBiometric();
-
-
-			if (biodata.getFaceHash() == null && biodata.getFingerHash() == null && biodata.getIris() == null)
-				return new MDSRCaptureModel();
-			} catch (Throwable t) {
-				logger.error(" Port issue " + contextKey, t);
-				t.getStackTrace();
-				return null;
-			}
-
-
-			if (bioExceptions != null && !bioExceptions.isEmpty()) {
-				for (int modalityCount = 0; modalityCount < bioExceptions.size(); modalityCount++)
-					bioexceptionlist.add(bioExceptions.get(modalityCount).getSubType().toString());
-			}
-
-
-			try {
-				if ((filteredAttribs != null && filteredAttribs.contains("face")) && biodata.getRawFaceData() != null) {
-
-					List<MDSDevice> faceDevices = mds.getRegDeviceInfo(DataProviderConstants.MDS_DEVICE_TYPE_FACE);
-					MDSDevice faceDevice = faceDevices.get(0);
-
-					capture = mds.captureFromRegDevice(faceDevice, capture, DataProviderConstants.MDS_DEVICE_TYPE_FACE,
-							null, 60, faceDevice.getDeviceSubId().get(0), port, contextKey, null);
-				}
-			}
-
-			catch (Throwable t) {
-				logger.error(" Face get capture   fail" + contextKey, t);
-				t.getStackTrace();
-				return null;
-			}
-
-
-			try {
-				if (biodata.getIris() != null) {
-					List<BioModality> irisExceptions = null;
-					List<String> listexceptionBio = new ArrayList<String>();
-
-				if (bioExceptions != null && !bioExceptions.isEmpty()) {
-					irisExceptions = getModalitiesByType(bioExceptions, "Iris");
-					for (BioModality modality : bioExceptions) {
-						listexceptionBio.add(modality.getSubType());
-
-					}
-				}
-
-				List<MDSDevice> irisDevices = mds.getRegDeviceInfo(DataProviderConstants.MDS_DEVICE_TYPE_IRIS);
-				MDSDevice irisDevice = irisDevices.get(0);
-
-				if (irisExceptions == null || irisExceptions.isEmpty()) {
-					if (filteredAttribs != null && filteredAttribs.contains(LEFTEYE)) {
-						capture = mds.captureFromRegDevice(irisDevice, capture,
-								DataProviderConstants.MDS_DEVICE_TYPE_IRIS, null, 60,
-								irisDevice.getDeviceSubId().get(0), port, contextKey, null);
-					}
-
-					if (irisDevice.getDeviceSubId().size() > 1) {
-						if (filteredAttribs != null && filteredAttribs.contains(RIGHTEYE)) {
-
-							capture = mds.captureFromRegDevice(irisDevice, capture,
-									DataProviderConstants.MDS_DEVICE_TYPE_IRIS, null, 60,
-									irisDevice.getDeviceSubId().get(1), port, contextKey, null);
-						}
-					}
-				} else {
-					String[] irisSubTypes = new String[irisExceptions.size()];
-					int i = 0;
-					for (BioModality bm : irisExceptions) {
-						irisSubTypes[i] = bm.getSubType();
-						i++;
-					}
-					for (String f : irisSubTypes) {
-
-						if (f.equalsIgnoreCase(RIGHT)
-								&& (filteredAttribs != null && filteredAttribs.contains(LEFTEYE))) {
-							capture = mds.captureFromRegDevice(irisDevice, capture,
-									DataProviderConstants.MDS_DEVICE_TYPE_IRIS, null, 60,
-									irisDevice.getDeviceSubId().get(0), port, contextKey, null);
-						} else if (f.equalsIgnoreCase("left")
-								&& (filteredAttribs != null && filteredAttribs.contains(RIGHTEYE))) {
-
-							if (irisDevice.getDeviceSubId().size() > 1)
-								capture = mds.captureFromRegDevice(irisDevice, capture,
-										DataProviderConstants.MDS_DEVICE_TYPE_IRIS, null, 60,
-										irisDevice.getDeviceSubId().get(1), port, contextKey, null);
-						}
-					}
-				}
-				}
-
-			}
-
-			catch (Throwable t) {
-				logger.error(" IRIS get capture  fail" + contextKey, t);
-				t.getStackTrace();
-				return null;
-			}
-
-			try {
-				if (biodata.getFingerPrint() != null) {
-				List<BioModality> fingerExceptions = null;
-				List<MDSDeviceCaptureModel> lstToRemove = new ArrayList<MDSDeviceCaptureModel>();
-				List<String> listFingerExceptionBio = new ArrayList<String>();
-
-				if (bioExceptions != null && !bioExceptions.isEmpty()) {
-
-					fingerExceptions = getModalitiesByType(bioExceptions, "Finger");
-
-					for (BioModality modality : bioExceptions) {
-						listFingerExceptionBio.add(modality.getSubType());
-
-					}
-				}
-
-				List<MDSDevice> fingerDevices = mds.getRegDeviceInfo(DataProviderConstants.MDS_DEVICE_TYPE_FINGER);
-				MDSDevice fingerDevice = fingerDevices.get(0);
-
-				for (int i = 0; i < fingerDevice.getDeviceSubId().size(); i++) {
-					capture = mds.captureFromRegDevice(fingerDevice, capture,
-							DataProviderConstants.MDS_DEVICE_TYPE_FINGER, null, 60,
-							fingerDevice.getDeviceSubId().get(i), port, contextKey, null);
-				}
-				List<MDSDeviceCaptureModel> lstFingers = capture.getLstBiometrics()
-						.get(DataProviderConstants.MDS_DEVICE_TYPE_FINGER);
-				if (filteredAttribs != null) {
-
-					String attr = null;
-
-					for (MDSDeviceCaptureModel mdc : lstFingers) {
-						int indx = 0;
-						boolean bFound = false;
-						for (indx = 0; indx < DataProviderConstants.schemaNames.length; indx++) {
-							if (DataProviderConstants.displayFingerName[indx].equals(mdc.getBioSubType())) {
-								attr = DataProviderConstants.schemaNames[indx];
-								break;
-							}
-						}
-						if (attr != null) {
-							for (String a : filteredAttribs) {
-								if (a.equals(attr)) {
-									bFound = true;
-									break;
-								}
-							}
-						}
-						if (!bFound)
-							lstToRemove.add(mdc);
-					}
-					lstFingers.removeAll(lstToRemove);
-				}
-
-				}
-			}
-
-			catch (Throwable t) {
-				logger.error("Finger get capture fail" + contextKey, t);
-				t.getStackTrace();
-				return null;
-			}
-
-			try {
-
-				if (bioExceptions != null && !bioExceptions.isEmpty()) {
-
-					List<MDSDevice> exceptionfaceDevices = mds
-							.getRegDeviceInfo(DataProviderConstants.MDS_DEVICE_TYPE_FACE);
-					MDSDevice exceptionDevice = exceptionfaceDevices.get(0);
-					try {
-						capture = mds.captureFromRegDevice(exceptionDevice, capture,
-								DataProviderConstants.MDS_DEVICE_TYPE_FACE, null, 60,
-								exceptionDevice.getDeviceSubId().get(0), port, contextKey, bioexceptionlist);
-
-					} catch (Throwable t) {
-						logger.error("Exception photo capture failure" + contextKey, t);
-						t.getStackTrace();
-						return null;
-					}
-
-				}
-
-			}
-
-			catch (Throwable t) {
-				logger.error("Exceptionphoto face capture", t);
-				t.getStackTrace();
-				return null;
-			}
-			return capture;
-		} finally {
-			cleanupMdsResources(mds, port, contextKey);
-		}
-	}
-
-	private static void cleanupMdsResources(MDSClientInterface mds, int port, String contextKey) {
-		try {
-			if (mds != null && port > 0) {
-				mds.setProfile("Default", port, contextKey);
-			}
-		} catch (Throwable t) {
-			logger.warn("Failed to reset MDS profile for context {}", contextKey, t);
-		} finally {
-			portmap.remove("port_" + contextKey);
-			try {
-				CentralizedMockSBI.stopSBI(contextKey);
-			} catch (Throwable t) {
-				logger.warn("Failed to stop SBI for context {}", contextKey, t);
-			}
-		}
+		return MdsCaptureService.regenBiometricViaMDS(resident, contextKey, purpose, qualityScore, process);
 	}
 
 	public static String toCBEFFFromCapture(List<String> bioFilter, MDSRCaptureModel capture, String toFile,
 			List<String> missAttribs, boolean genarateValidCbeff, List<BioModality> exceptionlist, String contextKey)
 			throws Exception {
-		logger.info("CAPTURE XML DATA: " + capture.toString());
-
-		String retXml = "";
-
-		String mosipVersion = null;
-		try {
-			mosipVersion = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "mosip.version").toString();
-		} catch (Exception e) {
-
-		}
-
-		XMLBuilder builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL)
-				.e(BIRINFO).e(INTEGRITY).t(FALSE).up().up();
-
-		builder.getDocument().setXmlStandalone(true);
-
-		List<String> bioSubType = new ArrayList<>();
-
-
-		try {
-			List<MDSDeviceCaptureModel> lstFingerData = capture.getLstBiometrics()
-					.get(DataProviderConstants.MDS_DEVICE_TYPE_FINGER);
-
-			builder = xmlbuilderFinger(bioFilter, lstFingerData, bioSubType, builder, exceptionlist, genarateValidCbeff,
-					contextKey);
-
-			if (exceptionlist != null && !exceptionlist.isEmpty()) {
-				builder = xmlbuilderFingerExep(bioFilter, exceptionlist, bioSubType, builder, genarateValidCbeff,
-						contextKey);
-			}
-		} catch (Exception e) {
-			logger.error("xmlbuilderFinger failed" + e.getMessage());
-		}
-
-
-		try {
-			if (bioFilter.contains("face")) {
-
-				List<MDSDeviceCaptureModel> lstFaceData = capture.getLstBiometrics()
-						.get(DataProviderConstants.MDS_DEVICE_TYPE_FACE);
-				bioSubType.add("face");
-				String faceXml = buildBirFace(lstFaceData.get(0).getBioValue(), lstFaceData.get(0).getSb(),
-						lstFaceData.get(0).getPayload(), lstFaceData.get(0).getQualityScore(), genarateValidCbeff,
-						FALSE, contextKey);
-				builder = builder.importXMLBuilder(XMLBuilder.parse(faceXml));
-
-			}
-		} catch (Exception e) {
-			logger.error("buildBirFace failed");
-			logger.error(e.getMessage());
-		}
-
-
-		try {
-			if (bioFilter.contains(LEFTEYE) && bioFilter.contains(RIGHTEYE)) {
-
-				List<MDSDeviceCaptureModel> lstIrisData = capture.getLstBiometrics()
-						.get(DataProviderConstants.MDS_DEVICE_TYPE_IRIS);
-				logger.info("IRIS DATA XML DATA: " + lstIrisData.toString());
-
-				builder = xmlbuilderIris(bioFilter, lstIrisData, bioSubType, builder, genarateValidCbeff, exceptionlist,
-						contextKey);
-			}
-			if (exceptionlist != null && !exceptionlist.isEmpty()) {
-				builder = xmlbuilderIrisExcep(bioFilter, exceptionlist, bioSubType, builder, genarateValidCbeff,
-						contextKey);
-			}
-		} catch (Exception e) {
-			logger.error("xmlbuilderIris failed");
-			logger.error(e.getMessage());
-		}
-
-
-		try {
-			if (exceptionlist != null && !exceptionlist.isEmpty()) {
-				List<MDSDeviceCaptureModel> lstFaceData = capture.getLstBiometrics().get(EXCEPTION);
-				bioSubType.add("exceptionphoto");
-				String faceXml = buildBirExceptionPhoto(lstFaceData.get(1).getBioValue(), lstFaceData.get(1).getSb(),
-						lstFaceData.get(1).getPayload(), lstFaceData.get(1).getQualityScore(), genarateValidCbeff,
-						FALSE, contextKey);
-				builder = builder.importXMLBuilder(XMLBuilder.parse(faceXml));
-			}
-		} catch (Exception e) {
-			logger.error("buildBirExceptionPhoto failed");
-			logger.error(e.getMessage());
-		}
-
-
-		if (mosipVersion != null && mosipVersion.startsWith("1.2") && !bioSubType.isEmpty()) {
-			builder.e(OTHERS).e("Key").t("CONFIGURED").up().e("Value")
-					.t(bioSubType.toString().substring(1, bioSubType.toString().length() - 1)).up().up();
-		}
-		if (toFile != null) {
-			FileOutputStream fos = new FileOutputStream(toFile);
-			PrintWriter writer = new PrintWriter(fos);
-			builder.toWriter(true, writer, null);
-			fos.close();
-		}
-
-		retXml = builder.asString(null);
-		return retXml;
+		return CbeffAssembler.toCBEFFFromCapture(bioFilter, capture, toFile, missAttribs, genarateValidCbeff,
+				exceptionlist, contextKey);
 	}
-
-	private static XMLBuilder xmlbuilderIris(List<String> bioFilter, List<MDSDeviceCaptureModel> lstIrisData,
-			List<String> bioSubType, XMLBuilder builder, boolean genarateValidCbeff, List<BioModality> exceptionlst,
-			String contextKey)
-
-	{
-		List<String> listWithoutExceptions = bioFilter;
-		if (exceptionlst != null && !exceptionlst.isEmpty()) {
-			List<String> exceptions = exceptionlst.stream().map(BioModality::getSubType).collect(Collectors.toList());
-			List<String> schemaName = new ArrayList<String>();
-			for (String ex : exceptions) {
-				schemaName.add(getschemaName(ex));
-			}
-			listWithoutExceptions = bioFilter.stream().filter(bioAttribute -> !schemaName.contains(bioAttribute))
-					.collect(Collectors.toList());
-		}
-		RestClient.logInfo(contextKey, "withoutExceptionList is: " + listWithoutExceptions);
-
-		try {
-			if (lstIrisData != null) {
-				String irisXml = "";
-				String leftQuality = null;
-				String rightQuality = null;
-				boolean leftPresent = false;
-				boolean rightPresent = false;
-
-				for (MDSDeviceCaptureModel cm : lstIrisData) {
-
-					if (listWithoutExceptions.contains(LEFTEYE) && cm.getBioSubType().equals("Left")) {
-						irisXml = buildBirIris(
-								cm.getBioValue(),
-								"Left",
-								cm.getSb(),
-								cm.getPayload(),
-								cm.getQualityScore(),
-								genarateValidCbeff,
-								FALSE,
-								contextKey);
-
-						builder = builder.importXMLBuilder(XMLBuilder.parse(irisXml));
-						bioSubType.add("Left");
-
-						leftQuality = cm.getQualityScore();
-						leftPresent = true;
-					} else if (!listWithoutExceptions.contains(LEFTEYE)) {
-						VariableManager.appendVariableValue(contextKey, "EXCEPTION_BIOMETRICS", "LE");
-
-					}
-
-					if (listWithoutExceptions.contains(RIGHTEYE) && cm.getBioSubType().equals(RIGHT)) {
-						irisXml = buildBirIris(
-								cm.getBioValue(),
-								RIGHT,
-								cm.getSb(),
-								cm.getPayload(),
-								cm.getQualityScore(),
-								genarateValidCbeff,
-								FALSE,
-								contextKey);
-
-						builder = builder.importXMLBuilder(XMLBuilder.parse(irisXml));
-						bioSubType.add(RIGHT);
-
-						rightQuality = cm.getQualityScore();
-						rightPresent = true;
-					} else if (!listWithoutExceptions.contains(RIGHTEYE)) {
-						VariableManager.appendVariableValue(contextKey, "EXCEPTION_BIOMETRICS", "RE");
-
-					}
-				}
-
-				String finalValue;
-
-				if (!leftPresent && !rightPresent) {
-					finalValue = "--Biometrics-Not-Available--";
-				} else if ((leftQuality != null && Double.parseDouble(leftQuality) >= 80)
-						|| (rightQuality != null && Double.parseDouble(rightQuality) >= 80)) {
-					finalValue = "level-9";
-				} else {
-					finalValue = "level-2";
-				}
-
-				VariableManager.setVariableValue(
-						contextKey,
-						"Biometric_Quality-Iris",
-						finalValue);
-			}
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-		}
-
-		return builder;
-	}
-
-	private static XMLBuilder xmlbuilderIrisExcep(List<String> bioFilter, List<BioModality> lstIrisData,
-			List<String> bioSubType, XMLBuilder builder, boolean genarateValidCbeff, String contextKey) {
-		try {
-			if (lstIrisData != null) {
-				for (BioModality finger : lstIrisData) {
-
-					if (!finger.getType().equalsIgnoreCase("Iris"))
-						continue;
-
-					String strFingerXml = buildBirIris(
-							finger.getType(),
-							finger.getSubType(),
-							Arrays.toString(new byte[0]),
-							"",
-							"0",
-							genarateValidCbeff,
-							"true",
-							contextKey);
-
-					XMLBuilder fbuilder = XMLBuilder.parse(strFingerXml);
-					builder = builder.importXMLBuilder(fbuilder);
-				}
-			}
-
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-		}
-
-		return builder;
-	}
-
-	private static String getschemaName(String name) {
-
-		for (int i = 0; i < 13; i++) {
-			String displayFingerName = DataProviderConstants.displayFullName[i];
-			if (displayFingerName.equalsIgnoreCase(name) == true)
-				return DataProviderConstants.schemaNames[i];
-		}
-
-
-		return name;
-	}
-
-	private static XMLBuilder xmlbuilderFinger(List<String> bioFilter, List<MDSDeviceCaptureModel> lstFingerData,
-			List<String> bioSubType, XMLBuilder builder, List<BioModality> exceptionlst, boolean genarateValidCbeff,
-			String contextKey) {
-
-		List<String> listWithoutExceptions = bioFilter;
-		if (exceptionlst != null && !exceptionlst.isEmpty()) {
-			List<String> exceptions = exceptionlst.stream().map(BioModality::getSubType).collect(Collectors.toList());
-			RestClient.logInfo(contextKey, "exceptions" + exceptions);
-			List<String> schemaName = new ArrayList<String>();
-			for (String ex : exceptions){
-				schemaName.add(getschemaName(ex));
-				VariableManager.appendVariableValue(contextKey, "EXCEPTION_BIOMETRICS", getFingerCode(ex));
-			}
-			listWithoutExceptions = bioFilter.stream().filter(bioAttribute -> !schemaName.contains(bioAttribute))
-					.collect(Collectors.toList());
-		}
-
-		int i = 0;
-		String fingerData = null;
-		boolean fingerPresent = false;
-		String maxQuality = null;
-
-		try {
-			if (lstFingerData != null) {
-				for (String finger : listWithoutExceptions) {
-
-					if (finger.toLowerCase().contains("eye") || finger.toLowerCase().equals("face"))
-						continue;
-
-					i = Arrays.asList(DataProviderConstants.schemaNames).indexOf(finger);
-					String displayName = DataProviderConstants.displayFingerName[i];
-					MDSDeviceCaptureModel currentCM = null;
-
-					for (MDSDeviceCaptureModel cm : lstFingerData) {
-						if (cm.getBioSubType().equals(displayName)) {
-							fingerData = cm.getBioValue();
-							bioSubType.add(finger);
-							currentCM = cm;
-							break;
-						}
-					}
-
-					RestClient.logInfo(contextKey, "fingerData is: " + fingerData);
-
-					if (i >= 0 && fingerData != null && currentCM != null) {
-
-						String strFinger = DataProviderConstants.displayFingerName[i];
-
-						String strFingerXml = buildBirFinger(
-								fingerData,
-								strFinger,
-								currentCM.getSb(),
-								currentCM.getPayload(),
-								currentCM.getQualityScore(),
-								genarateValidCbeff,
-								FALSE,
-								contextKey);
-
-						XMLBuilder fbuilder = XMLBuilder.parse(strFingerXml);
-						builder = builder.importXMLBuilder(fbuilder);
-
-						fingerPresent = true;
-
-						String q = currentCM.getQualityScore();
-
-						if (maxQuality == null || Double.parseDouble(q) > Double.parseDouble(maxQuality))
-							maxQuality = q;
-
-					} 
-				}
-
-				String finalValue;
-
-				if (!fingerPresent) {
-					finalValue = "--Biometrics-Not-Available--";
-				} else if (maxQuality != null && Double.parseDouble(maxQuality) >= 80) {
-					finalValue = "level-9";
-				} else {
-					finalValue = "level-2";
-				}
-
-				VariableManager.setVariableValue(contextKey, "Biometric_Quality-Finger", finalValue);
-			}
-
-		} catch (Exception e1) {
-			logger.error(e1.getMessage());
-		}
-
-		return builder;
-	}
-
-	private static String getFingerCode(String finger) {
-		if (finger == null)
-			return "";
-		String f = finger.toLowerCase().replace("finger:", "");
-		String side = f.contains("right") ? "R" : f.contains("left") ? "L" : "";
-		String type = "";
-		if (f.contains("thumb"))
-			type = "T";
-		else if (f.contains("index"))
-			type = "I";
-		else if (f.contains("middle"))
-			type = "M";
-		else if (f.contains("ring"))
-			type = "R";
-		else if (f.contains("little"))
-			type = "L";
-		if(type.isEmpty())
-			return "";
-		return side + type;
-	}
-
-	private static XMLBuilder xmlbuilderFingerExep(List<String> bioFilter, List<BioModality> lstFingerData,
-			List<String> bioSubType, XMLBuilder builder, boolean genarateValidCbeff, String contextKey)
-			throws ParserConfigurationException, FactoryConfigurationError, TransformerException, SAXException,
-			IOException {
-
-		if (lstFingerData != null) {
-			for (BioModality finger : lstFingerData) {
-				if (finger.getType().equalsIgnoreCase("iris") || finger.getType().equalsIgnoreCase("face"))
-					continue;
-
-				String strFingerXml = buildBirFinger(finger.getType(), finger.getSubType(),
-						Arrays.toString(new byte[0]), "", "0", genarateValidCbeff, "true", contextKey);
-				XMLBuilder fbuilder = XMLBuilder.parse(strFingerXml);
-				builder = builder.importXMLBuilder(fbuilder);
-			}
-
-		}
-
-		return builder;
-	}
-
 
 	public static String toCBEFF(List<String> bioFilter, BiometricDataModel biometricDataModel, String toFile,
 			boolean genarateValidCbeff, String contextKey) throws Exception {
-		String retXml = "";
-
-		XMLBuilder builder = XMLBuilder.create("BIR").a(XMLNS, XMLNS_URL)
-				.e(BIRINFO).e(INTEGRITY).t(FALSE).up().up();
-
-		builder.getDocument().setXmlStandalone(true);
-
-
-		String[] fingerPrint = biometricDataModel.getFingerPrint();
-
-
-		String qualityScore = null;
-		Hashtable<String, List<MDSDeviceCaptureModel>> capture = biometricDataModel.getCapture();
-		Enumeration<List<MDSDeviceCaptureModel>> elements = capture.elements();
-		while (elements.hasMoreElements()) {
-			List<MDSDeviceCaptureModel> nextElement = elements.nextElement();
-			qualityScore = nextElement.get(0).getQualityScore();
-			break;
-		}
-
-		int i = 0;
-		for (String finger : bioFilter) {
-			if (finger.toLowerCase().contains("eye") || finger.toLowerCase().equals("face"))
-				continue;
-			i = Arrays.asList(DataProviderConstants.schemaNames).indexOf(finger);
-
-			if (i >= 0) {
-				String strFinger = DataProviderConstants.displayFingerName[i];
-				String strFingerXml = buildBirFinger(fingerPrint[i], strFinger, null, null, qualityScore,
-						genarateValidCbeff, FALSE, contextKey);
-
-				XMLBuilder fbuilder = XMLBuilder.parse(strFingerXml);
-				builder = builder.importXMLBuilder(fbuilder);
-			}
-
-		}
-
-
-		if (bioFilter.contains("Face")) {
-			if (biometricDataModel.getEncodedPhoto() != null) {
-				String faceXml = buildBirFace(biometricDataModel.getEncodedPhoto(), null, null, qualityScore,
-						genarateValidCbeff, "true", contextKey);
-				builder = builder.importXMLBuilder(XMLBuilder.parse(faceXml));
-			}
-		}
-
-
-		IrisDataModel irisInfo = biometricDataModel.getIris();
-		if (irisInfo != null) {
-			String irisXml = "";
-			if (bioFilter.contains(LEFTEYE)) {
-				irisXml = buildBirIris(irisInfo.getLeft(), "Left", null, null, qualityScore, genarateValidCbeff, "true",
-						contextKey);
-				builder = builder.importXMLBuilder(XMLBuilder.parse(irisXml));
-			}
-			if (bioFilter.contains(RIGHTEYE)) {
-				irisXml = buildBirIris(irisInfo.getRight(), RIGHT, null, null, qualityScore, genarateValidCbeff, "true",
-						contextKey);
-				builder = builder.importXMLBuilder(XMLBuilder.parse(irisXml));
-			}
-		}
-
-		if (toFile != null) {
-			FileOutputStream fos = new FileOutputStream(toFile);
-			PrintWriter writer = new PrintWriter(fos);
-			builder.toWriter(true, writer, null);
-			fos.close();
-		}
-		retXml = builder.asString(null);
-		return retXml;
+		return CbeffAssembler.toCBEFF(bioFilter, biometricDataModel, toFile, genarateValidCbeff, contextKey);
 	}
 
 	public static Hashtable<Integer, List<File>> impressionCaptureList(String contextKey) {
-
-		String dirPath = System.getProperty("java.io.tmpdir")
-				+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.fingerprintdatapath").toString();
-		RestClient.logInfo(contextKey, DIRPATH + dirPath);
-		Hashtable<Integer, List<File>> tblFiles = new Hashtable<Integer, List<File>>();
-		File dir = new File(dirPath);
-
-		File listDir[] = dir.listFiles();
-		int numberOfSubfolders = listDir.length;
-
-		int min = 1;
-		int max = numberOfSubfolders;
-		int randomNumber = (int) (Math.random() * (max - min)) + min;
-		String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-		String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-		if (afterscenario.contains("_")) {
-			afterscenario = afterscenario.replace("_", "0");
-		}
-		int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-		int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-		RestClient.logInfo(contextKey, "currentScenarioNumber=" + currentScenarioNumber + " numberOfSubfolders="
-				+ numberOfSubfolders + " impressionToPick=" + impressionToPick);
-		List<File> lst = new LinkedList<File>();
-		lst = CommonUtil.listFiles(dirPath + String.format("/Impression_%d/fp_1/", impressionToPick));
-		tblFiles.put(impressionToPick, lst);
-		return tblFiles;
+		return FileBiometricLoader.impressionCaptureList(contextKey);
 	}
 
 	public static BiometricDataModel getBiometricData(Boolean bFinger, String contextKey) throws Exception {
-
-		BiometricDataModel data = new BiometricDataModel();
-
-		File tmpDir;
-
-		if (bFinger) {
-
-			Object val = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "enableExternalBiometricSource");
-			boolean bExternalSrc = false;
-			if (val != null)
-				bExternalSrc = Boolean.valueOf(val.toString());
-			if (bExternalSrc) {
-
-				String bioSrc = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "externalBiometricsource")
-						.toString();
-
-				String[] fingerPrints = new String[10];
-				String[] fingerPrintHash = new String[10];
-				byte[][] fingerPrintRaw = new byte[10][1];
-
-				for (int i = 0; i < 10; i++) {
-					String modalityName = DataProviderConstants.MDSProfileFingerNames[i];
-					modalityName = modalityName.replace('_', ' ');
-					String fPath = bioSrc + modalityName + ".jp2";
-					byte[] fdata = CommonUtil.read(fPath);
-					fingerPrintRaw[i] = fdata;
-					fingerPrints[i] = Base64.getEncoder().encodeToString(fdata);
-					try {
-						fingerPrintHash[i] = CommonUtil.getHexEncodedHash(fdata);
-
-					} catch (Exception e) {
-						logger.error(e.getMessage());
-					}
-					data.setFingerPrint(fingerPrints);
-					data.setFingerHash(fingerPrintHash);
-					data.setFingerRaw(fingerPrintRaw);
-
-				}
-
-				return data;
-			}
-
-			Boolean bAnguli = Boolean.parseBoolean(
-					VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "enableAnguli").toString());
-			if (bAnguli) {
-
-
-				try {
-					tmpDir = Files.createTempDirectory("fps").toFile();
-					Hashtable<Integer, List<File>> prints = generateFingerprint(tmpDir.getAbsolutePath(), 10, 2, 4,
-							FPClassDistribution.arch, contextKey);
-					List<File> firstSet = prints.get(1);
-
-					String[] fingerPrints = new String[10];
-					String[] fingerPrintHash = new String[10];
-					byte[][] fingerPrintRaw = new byte[10][1];
-
-					int index = 0;
-					for (File f : firstSet) {
-
-						if (index > 9)
-							break;
-
-						byte[] fdata = CommonUtil.read(f.getAbsolutePath());
-						fingerPrintRaw[index] = fdata;
-						fingerPrints[index] = Base64.getEncoder().encodeToString(fdata);
-
-
-						try {
-							fingerPrintHash[index] = CommonUtil.getHexEncodedHash(fdata);
-
-						} catch (Exception e) {
-							logger.error(e.getMessage());
-						}
-
-						index++;
-
-					}
-					data.setFingerPrint(fingerPrints);
-					data.setFingerHash(fingerPrintHash);
-					data.setFingerRaw(fingerPrintRaw);
-					tmpDir.deleteOnExit();
-
-				} catch (IOException e) {
-					logger.error(e.getMessage());
-				}
-			} else {
-
-				String dirPath = System.getProperty("java.io.tmpdir") + VariableManager
-						.getVariableValue(contextKey, "mosip.test.persona.fingerprintdatapath").toString();
-				RestClient.logInfo(contextKey, DIRPATH + dirPath);
-				Hashtable<Integer, List<File>> tblFiles = new Hashtable<Integer, List<File>>();
-				File dir = new File(dirPath);
-
-				File listDir[] = null;
-				if (dir.isDirectory()) {
-
-					listDir = dir.listFiles(new FileFilter() {
-						@Override
-						public boolean accept(File file) {
-
-							return file.isDirectory() && file.getName().startsWith("Impression");
-						}
-					});
-				} else {
-					logger.error(dirPath + " is not a directory.");
-				}
-				int numberOfSubfolders = listDir.length;
-
-				int min = 1;
-				int max = numberOfSubfolders;
-				int randomNumber = (int) (Math.random() * (max - min)) + min;
-				String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-				String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-				if (afterscenario.contains("_")) {
-					afterscenario = afterscenario.replace("_", "0");
-				}
-				int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-				int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber
-						: randomNumber;
-				dirPath = FingerprintVariationGenerator.fingerprintVariationGenerator(contextKey, currentScenarioNumber,
-						impressionToPick);
-
-				RestClient.logInfo(contextKey, "currentScenarioNumber=" + currentScenarioNumber + " numberOfSubfolders="
-						+ numberOfSubfolders + " impressionToPick=" + impressionToPick);
-
-				List<File> firstSet = CommonUtil.listFiles(dirPath + "/fp/");
-
-				String[] fingerPrints = new String[10];
-				String[] fingerPrintHash = new String[10];
-				byte[][] fingerPrintRaw = new byte[10][1];
-				RestClient.logInfo(contextKey, "Impression used " + impressionToPick);
-
-				int index = 0;
-				for (File f : firstSet) {
-
-					if (index > 9)
-						break;
-
-					byte[] fdata;
-					try {
-						fdata = CommonUtil.read(f.getAbsolutePath());
-						fingerPrintRaw[index] = fdata;
-						fingerPrints[index] = Base64.getEncoder().encodeToString(fdata);
-
-						fingerPrintHash[index] = CommonUtil.getHexEncodedHash(fdata);
-
-					} catch (Exception e) {
-						logger.error(e.getMessage());
-					}
-					index++;
-
-				}
-				data.setFingerPrint(fingerPrints);
-				data.setFingerHash(fingerPrintHash);
-				data.setFingerRaw(fingerPrintRaw);
-				CommonUtil.deleteOldTempDir(dirPath);
-			}
-
-		}
-
-		return data;
+		return FileBiometricLoader.getBiometricData(bFinger, contextKey);
 	}
 
 	public static BiometricDataModel updateFingerData(String contextKey) throws Exception {
-
-		BiometricDataModel data = new BiometricDataModel();
-
-		String dirPath = System.getProperty("java.io.tmpdir")
-				+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.fingerprintdatapath").toString();
-		RestClient.logInfo(contextKey, DIRPATH + dirPath);
-		Hashtable<Integer, List<File>> tblFiles = new Hashtable<Integer, List<File>>();
-		File dir = new File(dirPath);
-
-		File listDir[] = null;
-		if (dir.isDirectory()) {
-
-			listDir = dir.listFiles(new FileFilter() {
-				@Override
-				public boolean accept(File file) {
-
-					return file.isDirectory() && file.getName().startsWith("Impression");
-				}
-			});
-		} else {
-			logger.error(dirPath + " is not a directory.");
-		}
-		int numberOfSubfolders = listDir.length;
-
-		int min = 1;
-		int max = numberOfSubfolders;
-		int randomNumber;
-		String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-		String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-		if (afterscenario.contains("_")) {
-			afterscenario = afterscenario.replace("_", "0");
-		}
-		int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-		randomNumber = (int) (Math.random() * (max - min)) + min;
-		int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-		dirPath = FingerprintVariationGenerator.fingerprintVariationGenerator(contextKey, currentScenarioNumber,
-				impressionToPick);
-
-		List<File> firstSet = CommonUtil.listFiles(dirPath + "/fp/");
-
-		String[] fingerPrints = new String[10];
-		String[] fingerPrintHash = new String[10];
-		byte[][] fingerPrintRaw = new byte[10][1];
-		RestClient.logInfo(contextKey, "Impression used " + impressionToPick);
-
-		int index = 0;
-		for (File f : firstSet) {
-
-			if (index > 9)
-				break;
-
-			byte[] fdata;
-			try {
-				fdata = CommonUtil.read(f.getAbsolutePath());
-				fingerPrintRaw[index] = fdata;
-				fingerPrints[index] = Base64.getEncoder().encodeToString(fdata);
-
-				fingerPrintHash[index] = CommonUtil.getHexEncodedHash(fdata);
-
-			} catch (Exception e) {
-				logger.error(e.getMessage());
-			}
-			index++;
-
-		}
-		data.setFingerPrint(fingerPrints);
-		data.setFingerHash(fingerPrintHash);
-		data.setFingerRaw(fingerPrintRaw);
-		CommonUtil.deleteOldTempDir(dirPath);
-
-		return data;
+		return FileBiometricLoader.updateFingerData(contextKey);
 	}
-
-	private static final Map<String, Integer> FINGER_INDEX_MAP = Map.of("rightindex", 0, "rightmiddle", 1, "rightring",
-			2, "rightlittle", 3, "rightthumb", 4, "leftindex", 5, "leftmiddle", 6, "leftring", 7, "leftlittle", 8,
-			"leftthumb", 9);
 
 	public static BiometricDataModel updateSelectedFingerData(ResidentModel model, String contextKey,
 			String fingerArgument) throws Exception {
-
-
-		BiometricDataModel data = model.getBiometric();
-
-
-		String[] fingerPrints = data.getFingerPrint();
-		String[] fingerPrintHash = data.getFingerHash();
-		byte[][] fingerPrintRaw = data.getFingerRaw();
-
-
-		if (fingerPrints == null)
-			fingerPrints = new String[10];
-		if (fingerPrintHash == null)
-			fingerPrintHash = new String[10];
-		if (fingerPrintRaw == null)
-			fingerPrintRaw = new byte[10][1];
-
-		String dirPath = System.getProperty("java.io.tmpdir")
-				+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.fingerprintdatapath").toString();
-
-		RestClient.logInfo(contextKey, DIRPATH + dirPath);
-
-		File dir = new File(dirPath);
-		if (!dir.isDirectory()) {
-			logger.error(dirPath + " is not a directory.");
-			return data;
-		}
-
-		File[] listDir = dir.listFiles(f -> f.isDirectory() && f.getName().startsWith("Impression"));
-		if (listDir == null || listDir.length == 0) {
-			logger.error("No impression directories found in {}", dirPath);
-			return data;
-		}
-		int numberOfSubfolders = listDir.length;
-		int min = 1;
-		int max = numberOfSubfolders;
-		int randomNumber;
-
-		String beforeScenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-
-		String afterScenario = beforeScenario.substring(0, beforeScenario.indexOf(':')).replace("_", "0");
-
-		int currentScenarioNumber = Integer.parseInt(afterScenario);
-
-		randomNumber = (int) (Math.random() * (max - min)) + min;
-		int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-		RestClient.logInfo(contextKey, "Impression used " + impressionToPick);
-
-		dirPath = FingerprintVariationGenerator.fingerprintVariationGenerator(contextKey, currentScenarioNumber,
-				impressionToPick);
-
-		List<File> firstSet = CommonUtil.listFiles(dirPath + "/fp/");
-		Collections.sort(firstSet, Comparator.comparing(File::getName));
-
-
-		Set<Integer> indexesToUpdate = Arrays.stream(fingerArgument.split(",")).map(String::trim)
-				.map(String::toLowerCase).map(FINGER_INDEX_MAP::get).filter(Objects::nonNull)
-				.collect(Collectors.toSet());
-
-
-		for (Integer index : indexesToUpdate) {
-			if (index >= firstSet.size()) {
-				logger.warn("Fingerprint file not found for index {}", index);
-				continue;
-			}
-
-			try {
-				File f = firstSet.get(index);
-				byte[] fdata = CommonUtil.read(f.getAbsolutePath());
-
-				fingerPrintRaw[index] = fdata;
-				fingerPrints[index] = Base64.getEncoder().encodeToString(fdata);
-				fingerPrintHash[index] = CommonUtil.getHexEncodedHash(fdata);
-
-				RestClient.logInfo(contextKey, "Updated finger index " + index);
-
-			} catch (Exception e) {
-				logger.error("Error updating finger index " + index, e);
-			}
-		}
-
-
-		data.setFingerPrint(fingerPrints);
-		data.setFingerHash(fingerPrintHash);
-		data.setFingerRaw(fingerPrintRaw);
-
-		CommonUtil.deleteOldTempDir(dirPath);
-		return data;
-	}
-
-
-	static Hashtable<Integer, List<File>> generateFingerprint(String outDir, int nFingerPrints,
-			int nImpressionsPerPrints, int nThreads, FPClassDistribution classDist, String contextKey) {
-
-		Hashtable<Integer, List<File>> tblFiles = new Hashtable<Integer, List<File>>();
-
-
-		String[] commands = { DataProviderConstants.ANGULI_PATH + "/Anguli.exe", "-outdir", outDir, "-numT",
-				String.format("%d", nThreads), "-num", String.format("%d", nFingerPrints), "-ni",
-				String.format("%d", nImpressionsPerPrints), "-cdist", classDist.name() };
-		RestClient.logInfo(contextKey, "Anguli commands" + commands);
-		ProcessBuilder pb = new ProcessBuilder(commands);
-		pb.directory(new File(DataProviderConstants.ANGULI_PATH));
-		pb.redirectErrorStream(true);
-
-		try {
-			Process proc = pb.start(); 
-			try (BufferedReader processOutput = new BufferedReader(new InputStreamReader(proc.getInputStream()))) {
-				String s;
-				while ((s = processOutput.readLine()) != null) {
-					RestClient.logInfo(contextKey, s);
-				}
-				int exitCode = proc.waitFor();
-				if (exitCode != 0) {
-					logger.warn("Anguli exited with code {} for context {}", exitCode, contextKey);
-				}
-				for (int i = 1; i <= nImpressionsPerPrints; i++) {
-
-					List<File> lst = CommonUtil.listFiles(outDir + String.format("/Impression_%d/fp_1/", i));
-					tblFiles.put(i, lst);
-				}
-			} catch (InterruptedException e) {
-				Thread.currentThread().interrupt();
-				logger.error("Anguli execution interrupted for context {}", contextKey, e);
-			} finally {
-				proc.destroy();
-			}
-		} catch (IOException e) {
-			logger.error(e.getMessage());
-		}
-		return tblFiles;
+		return FileBiometricLoader.updateSelectedFingerData(model, contextKey, fingerArgument);
 	}
 
 	public static IrisDataModel loadIris(String filePath, String subModality, IrisDataModel im) throws Exception {
-
-		IrisDataModel m = im;
-		if (m == null)
-			m = new IrisDataModel();
-		String irisData = "";
-		String irisHash = "";
-
-		if (Files.exists(Paths.get(filePath))) {
-			byte[] fdata = CommonUtil.read(filePath);
-			irisData = Hex.encodeHexString(fdata);
-			irisHash = CommonUtil.getHexEncodedHash(fdata);
-			if (subModality.equals("left")) {
-				m.setLeftHash(irisHash);
-				m.setLeft(irisData);
-			} else {
-				m.setRightHash(irisHash);
-				m.setRight(irisData);
-			}
-		}
-
-		return m;
+		return FileBiometricLoader.loadIris(filePath, subModality, im);
 	}
-
 
 	public static List<IrisDataModel> generateIris(int count, String contextKey) throws Exception {
-		List<IrisDataModel> retVal = new ArrayList<IrisDataModel>();
-
-		IrisDataModel m = new IrisDataModel();
-
-		Object val = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "enableExternalBiometricSource");
-		boolean bExternalSrc = false;
-
-
-		if (val != null)
-			bExternalSrc = Boolean.valueOf(val.toString());
-
-		if (bExternalSrc) {
-
-
-			String bioSrc = VariableManager.getVariableValue(VariableManager.NS_DEFAULT, "externalBiometricsource")
-					.toString();
-			String fPathL = bioSrc + "Left Iris.jp2";
-			String fPathR = bioSrc + "Right Iris.jp2";
-
-			String leftIrisData = "";
-			String rightIrisData = "";
-			String irisHash = "";
-			byte[] fdata = null;
-			byte[] frdata = null;
-			if (Files.exists(Paths.get(fPathL))) {
-				fdata = CommonUtil.read(fPathL);
-				leftIrisData = Hex.encodeHexString(fdata);
-				irisHash = CommonUtil.getHexEncodedHash(fdata);
-				m.setLeftHash(irisHash);
-			}
-			if (Files.exists(Paths.get(fPathR))) {
-				frdata = CommonUtil.read(fPathR);
-				rightIrisData = Hex.encodeHexString(frdata);
-				irisHash = CommonUtil.getHexEncodedHash(frdata);
-				m.setRightHash(irisHash);
-			}
-			if (leftIrisData.equals("")) {
-				fdata = frdata;
-				leftIrisData = rightIrisData;
-			} else if (rightIrisData.equals("")) {
-				frdata = fdata;
-				rightIrisData = leftIrisData;
-			}
-			m.setLeft(leftIrisData);
-			m.setRight(rightIrisData);
-			m.setRawLeft(fdata);
-			m.setRawRight(frdata);
-			retVal.add(m);
-		} else {
-
-			String srcPath = System.getProperty("java.io.tmpdir")
-					+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.irisdatapath").toString();
-
-			int num = new File(srcPath).list().length;
-			int[] index = CommonUtil.generateRandomNumbers(count, num, 1);
-			String leftbmp = null;
-			String rightbmp = null;
-
-			RestClient.logInfo(contextKey, DIRPATH + srcPath);
-			Hashtable<Integer, List<File>> tblFiles = new Hashtable<Integer, List<File>>();
-			File dir = new File(srcPath);
-
-			File listDir[] = dir.listFiles();
-			int numberOfSubfolders = listDir.length;
-
-			int min = 1;
-			int max = numberOfSubfolders;
-			int randomNumber = (int) (Math.random() * (max - min)) + min;
-			String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-			String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-			if (afterscenario.contains("_")) {
-				afterscenario = afterscenario.replace("_", "0");
-			}
-
-			int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-			int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-			srcPath = IrisVariationGenerator.irisVariationGenerator(contextKey, currentScenarioNumber,
-					impressionToPick);
-
-			File folder = new File(srcPath + "/iris/");
-			logger.info(srcPath + "/iris/");
-
-			File[] listOfFiles = folder.listFiles();
-
-
-			for (File file : listOfFiles) {
-				if (file.getName().contains("L")) {
-					leftbmp = file.getName();
-				} else {
-					rightbmp = file.getName();
-				}
-			}
-
-			if (leftbmp == null) {
-				leftbmp = rightbmp;
-			}
-			if (rightbmp == null) {
-				rightbmp = leftbmp;
-			}
-			String fPathL = srcPath + "/" + "iris" + "/" + leftbmp;
-			String fPathR = srcPath + "/" + "iris" + "/" + rightbmp;
-			String leftIrisData = "";
-			String rightIrisData = "";
-			String irisHash = "";
-			byte[] fldata = null;
-			byte[] frdata = null;
-			if (Files.exists(Paths.get(fPathL))) {
-				fldata = CommonUtil.read(fPathL);
-				leftIrisData = Hex.encodeHexString(fldata);
-				irisHash = CommonUtil.getHexEncodedHash(fldata);
-				m.setLeftHash(irisHash);
-			}
-			if (Files.exists(Paths.get(fPathR))) {
-				frdata = CommonUtil.read(fPathR);
-				rightIrisData = Hex.encodeHexString(frdata);
-				irisHash = CommonUtil.getHexEncodedHash(frdata);
-				m.setRightHash(irisHash);
-			}
-			if (leftIrisData.equals("")) {
-				fldata = frdata;
-				leftIrisData = rightIrisData;
-			} else if (rightIrisData.equals("")) {
-				frdata = fldata;
-				rightIrisData = leftIrisData;
-			}
-			m.setLeft(leftIrisData);
-			m.setRight(rightIrisData);
-			m.setRawLeft(fldata);
-			m.setRawRight(frdata);
-			retVal.add(m);
-			CommonUtil.deleteOldTempDir(srcPath);
-		}
-
-		return retVal;
+		return FileBiometricLoader.generateIris(count, contextKey);
 	}
 
-
 	static List<IrisDataModel> updateIris(String contextKey) throws Exception {
-
-		List<IrisDataModel> retVal = new ArrayList<IrisDataModel>();
-		IrisDataModel m = new IrisDataModel();
-		String srcPath = System.getProperty("java.io.tmpdir")
-				+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.irisdatapath").toString();
-		String leftbmp = null;
-		String rightbmp = null;
-
-		RestClient.logInfo(contextKey, DIRPATH + srcPath);
-		File dir = new File(srcPath);
-
-		File listDir[] = dir.listFiles();
-		int numberOfSubfolders = listDir.length;
-
-		int min = 1;
-		int max = numberOfSubfolders;
-		int randomNumber;
-		String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-		String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-		if (afterscenario.contains("_")) {
-			afterscenario = afterscenario.replace("_", "0");
-		}
-		int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-		randomNumber = (int) (Math.random() * (max - min)) + min;
-		int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-		srcPath = IrisVariationGenerator.irisVariationGenerator(contextKey, currentScenarioNumber, impressionToPick);
-
-		File folder = new File(srcPath + "/iris/");
-
-		File[] listOfFiles = folder.listFiles();
-
-		for (File file : listOfFiles) {
-			if (file.getName().contains("L")) {
-				leftbmp = file.getName();
-			} else {
-				rightbmp = file.getName();
-			}
-		}
-
-		if (leftbmp == null) {
-			leftbmp = rightbmp;
-		}
-		if (rightbmp == null) {
-			rightbmp = leftbmp;
-		}
-		String fPathL = srcPath + "/iris/" + leftbmp;
-		String fPathR = srcPath + "/iris/" + rightbmp;
-
-		String leftIrisData = "";
-		String rightIrisData = "";
-		String irisHash = "";
-		byte[] fldata = null;
-		byte[] frdata = null;
-		if (Files.exists(Paths.get(fPathL))) {
-			fldata = CommonUtil.read(fPathL);
-			leftIrisData = Hex.encodeHexString(fldata);
-			irisHash = CommonUtil.getHexEncodedHash(fldata);
-			m.setLeftHash(irisHash);
-		}
-		if (Files.exists(Paths.get(fPathR))) {
-			frdata = CommonUtil.read(fPathR);
-			rightIrisData = Hex.encodeHexString(frdata);
-			irisHash = CommonUtil.getHexEncodedHash(frdata);
-			m.setRightHash(irisHash);
-		}
-		if (leftIrisData.equals("")) {
-			fldata = frdata;
-			leftIrisData = rightIrisData;
-		} else if (rightIrisData.equals("")) {
-			frdata = fldata;
-			rightIrisData = leftIrisData;
-		}
-		m.setLeft(leftIrisData);
-		m.setRight(rightIrisData);
-		m.setRawLeft(fldata);
-		m.setRawRight(frdata);
-		retVal.add(m);
-		CommonUtil.deleteOldTempDir(srcPath);
-		return retVal;
+		return FileBiometricLoader.updateIris(contextKey);
 	}
 
 	static List<IrisDataModel> updateSelectedIris(ResidentModel model, String contextKey, String irisArgument)
 			throws Exception {
-
-		List<IrisDataModel> retVal = new ArrayList<>();
-
-
-		IrisDataModel m = model.getBiometric().getIris();
-		if (m == null) {
-			m = new IrisDataModel();
-		}
-
-		String srcPath = System.getProperty("java.io.tmpdir")
-				+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.irisdatapath").toString();
-
-		RestClient.logInfo(contextKey, DIRPATH + srcPath);
-
-		File dir = new File(srcPath);
-		File[] listDir = dir.listFiles();
-		if (listDir == null || listDir.length == 0) {
-			logger.error("No subdirectories found in {}", srcPath);
-			retVal.add(m);
-			return retVal;
-		}
-		int numberOfSubfolders = listDir.length;
-		int min = 1;
-		int max = numberOfSubfolders;
-		int randomNumber;
-
-		String beforeScenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-
-		String afterScenario = beforeScenario.substring(0, beforeScenario.indexOf(':')).replace("_", "0");
-
-		int currentScenarioNumber = Integer.parseInt(afterScenario);
-
-		randomNumber = (int) (Math.random() * (max - min)) + min;
-		int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-		srcPath = IrisVariationGenerator.irisVariationGenerator(contextKey, currentScenarioNumber, impressionToPick);
-
-		File folder = new File(srcPath + "/iris/");
-
-		File[] listOfFiles = folder.listFiles();
-		if (listOfFiles == null || listOfFiles.length == 0) {
-			logger.warn("No iris files found in folder {}", folder.getAbsolutePath());
-			retVal.add(m);
-			return retVal;
-		}
-		String leftbmp = null;
-		String rightbmp = null;
-
-		for (File file : listOfFiles) {
-			if (file.getName().contains("L")) {
-				leftbmp = file.getName();
-			} else {
-				rightbmp = file.getName();
-			}
-		}
-
-		String fPathL = leftbmp == null ? null
-				: srcPath + "/iris/" + leftbmp;
-
-		String fPathR = rightbmp == null ? null
-				: srcPath + "/iris/" + rightbmp;
-
-
-		Set<String> irisToUpdate = Arrays.stream(irisArgument.split(",")).map(String::trim).map(String::toLowerCase)
-				.collect(Collectors.toSet());
-
-
-		if (irisToUpdate.contains("leftiris") && fPathL != null && Files.exists(Paths.get(fPathL))) {
-
-			byte[] fldata = CommonUtil.read(fPathL);
-			m.setLeft(Hex.encodeHexString(fldata));
-			m.setLeftHash(CommonUtil.getHexEncodedHash(fldata));
-			m.setRawLeft(fldata);
-		}
-
-
-		if (irisToUpdate.contains("rightiris") && fPathR != null && Files.exists(Paths.get(fPathR))) {
-
-			byte[] frdata = CommonUtil.read(fPathR);
-			m.setRight(Hex.encodeHexString(frdata));
-			m.setRightHash(CommonUtil.getHexEncodedHash(frdata));
-			m.setRawRight(frdata);
-		}
-
-		retVal.add(m);
-		CommonUtil.deleteOldTempDir(srcPath);
-		return retVal;
+		return FileBiometricLoader.updateSelectedIris(model, contextKey, irisArgument);
 	}
 
 	static byte[][] updateFaceData(String contextKey) {
+		return FileBiometricLoader.updateFaceData(contextKey);
+	}
 
-		byte[] bencoded = null;
-		byte[] bData = null;
-		try {
-
-			String dirPath = System.getProperty("java.io.tmpdir")
-					+ VariableManager.getVariableValue(contextKey, "mosip.test.persona.facedatapath").toString();
-
-			File dir = new File(dirPath);
-			FileFilter filter = new FileFilter() {
-				@Override
-				public boolean accept(File pathname) {
-					return !pathname.isDirectory();
-				}
-			};
-			File[] listDir = dir.listFiles(filter);
-			int numberOfSubfolders = listDir.length;
-
-			int min = 1;
-			int max = numberOfSubfolders;
-			int randomNumber;
-			String beforescenario = VariableManager.getVariableValue(contextKey, SCENARIO).toString();
-			String afterscenario = beforescenario.substring(0, beforescenario.indexOf(':'));
-			if (afterscenario.contains("_")) {
-				afterscenario = afterscenario.replace("_", "0");
-			}
-			int currentScenarioNumber = Integer.valueOf(afterscenario);
-
-
-			randomNumber = (int) (Math.random() * (max - min)) + min;
-			int impressionToPick = (currentScenarioNumber < numberOfSubfolders) ? currentScenarioNumber : randomNumber;
-
-			dirPath = FaceVariationGenerator.faceVariationGenerator(contextKey, currentScenarioNumber,
-					impressionToPick);
-
-			List<File> firstSet = CommonUtil.listFiles(dirPath + "/face/");
-
-			BufferedImage img = null;
-
-			try (FileInputStream fos = new FileInputStream(firstSet.get(0));
-					BufferedInputStream bis = new BufferedInputStream(fos)) {
-				img = ImageIO.read(bis);
-				logger.info("Image picked from this path=" + firstSet.get(0));
-			}
-			ByteArrayOutputStream baos = new ByteArrayOutputStream();
-
-			ImageIO.write(img, "jpg", baos);
-			baos.flush();
-			bData = baos.toByteArray();
-			bencoded = PhotoProvider.encodeFaceImageData(bData);
-
-			baos.close();
-			CommonUtil.deleteOldTempDir(dirPath);
-		} catch (Exception e) {
-
-			logger.error(e.getMessage());
-		}
-		return new byte[][] { bencoded, bData };
+	public static File[] getRandomIrisVariation(File[] listOfFiles) {
+		return FileBiometricLoader.getRandomIrisVariation(listOfFiles);
 	}
 
 	public static void main(String[] args) throws Exception {
 
 		try {
 
-			String value = buildBirFinger("addfdfd", "finger", "jwtSign", PAYLOAD, null, true, FALSE, null);
+			String value = CbeffBirBuilder.buildBirFinger("addfdfd", "finger", "jwtSign", PAYLOAD, null, true, FALSE,
+					null);
 			logger.info(value);
 		} catch (FileNotFoundException e2) {
 			logger.error(e2.getMessage());
@@ -1845,82 +145,5 @@ public class BiometricDataProvider {
 			logger.error(e.getMessage());
 		}
 		logger.info(xml);
-	}
-
-	public static File[] getRandomIrisVariation(File[] listOfFiles) {
-		if (listOfFiles == null || listOfFiles.length == 0) {
-			System.out.println("No image files found in the directory.");
-			return null;
-		}
-
-
-		Map<String, File> leftIrisImages = new HashMap<>();
-		Map<String, File> rightIrisImages = new HashMap<>();
-
-
-		for (File file : listOfFiles) {
-			String fileName = file.getName();
-			if (fileName.endsWith("_L.bmp")) {
-				String baseName = fileName.replaceAll("_[0-9]+_L\\.bmp", "").replace("_L.bmp", "");
-				leftIrisImages.put(baseName, file);
-			} else if (fileName.endsWith("_R.bmp")) {
-				String baseName = fileName.replaceAll("_[0-9]+_R\\.bmp", "").replace("_R.bmp", "");
-				rightIrisImages.put(baseName, file);
-			}
-		}
-
-
-		List<String> commonNames = new ArrayList<>(leftIrisImages.keySet());
-		commonNames.retainAll(rightIrisImages.keySet());
-
-		if (commonNames.isEmpty()) {
-			System.out.println("No matching left and right iris images found.");
-			return null;
-		}
-
-
-		Random random = new Random();
-		String selectedName = commonNames.get(random.nextInt(commonNames.size()));
-
-		File selectedLeftImage = leftIrisImages.get(selectedName);
-		File selectedRightImage = rightIrisImages.get(selectedName);
-
-
-		File[] selectedImages = { selectedLeftImage, selectedRightImage };
-
-
-		System.out.println("Selected Left Iris Image: " + selectedLeftImage.getAbsolutePath());
-		System.out.println("Selected Right Iris Image: " + selectedRightImage.getAbsolutePath());
-
-		return selectedImages;
-	}
-
-	static void copyExceptionPhoto(
-			String basePath,
-			String sourceFolderName,
-			String targetFolderName) throws IOException {
-
-		Path sourceFile = Paths.get(
-				basePath,
-				sourceFolderName + "/Registration/",
-				"Exception_Photo.iso");
-
-		Path targetDir = Paths.get(basePath, targetFolderName + "/Registration/");
-		Path targetFile = targetDir.resolve("Exception_Photo.iso");
-
-		if (!Files.exists(sourceFile)) {
-			throw new IllegalArgumentException(
-					"Exception_Photo.iso not found in source folder: " + sourceFile);
-		}
-
-
-		if (!Files.exists(targetDir)) {
-			Files.createDirectories(targetDir);
-		}
-
-		Files.copy(
-				sourceFile,
-				targetFile,
-				StandardCopyOption.REPLACE_EXISTING);
 	}
 }
